client/requirements.txt
elasticsearch==8.10.0
kafka-python==2.0.2
rich==13.6.0
